refactor: move and rewrite splash screen
