--- crates/alerion_core/src/logging.rs
@@ -1,10 +1,0 @@
-pub fn splash() {
-    println!(
-        "
- █████  ██      ███████ ██████  ██  ██████  ███    ██ 
-██   ██ ██      ██      ██   ██ ██ ██    ██ ████   ██ 
-███████ ██      █████   ██████  ██ ██    ██ ██ ██  ██ 
-██   ██ ██      ██      ██   ██ ██ ██    ██ ██  ██ ██ 
-██   ██ ███████ ███████ ██   ██ ██  ██████  ██   ████ "
-    );
-}
